Fix: in admin page, user are deleted even if admin disagree to delete

+++ projects/ngleaf/src/components/pages/admin-settings-page/admin-settings-users/admin-settings-users.component.ts
@@ -38,8 +38,10 @@ export class AdminSettingsUsersComponent implements OnInit {
       data: dialogData
     });
 
-    dialogRef.afterClosed().subscribe(dialogResult => {
-      this.adminService.deleteAccount(account.id);
+    dialogRef.afterClosed().subscribe(confirmed => {
+      if (confirmed) {
+        this.adminService.deleteAccount(account.id);
+      }
     });
   }
 }
